Add files via upload
Project: Playable Black Jack Game:

Requirements:
Blackjack Challenge We would like for you to build the game of Blackjack using the command line in Java. The game has some interesting rules and is a great parallel for various business rules we encounter in real projects. Our objective here is to see how you organize your thoughts and implement major principles into your work. We will also assess your overall coding skills in the Java language. Your objective it to produce a PLAYABLE GAME. This does not mean all the rules need to be perfect, but the general game play should be implemented.This should be a command-line program, you will be playing against a very rudimentary dealer robot that you will write. THE ROBOT DOES NOT NEED TO BE GOOD AT BLACKJACK. It can randomly select moves, but it must make those moves within the allowed rules of the game.The game should start at $100, and each hand is $10. There are many good guides for how to play Blackjack, and here is a recap of the gamehttp://www.pagat.com/banking/blackjack.htmlThough the rules can get involved, the available actions available are quite simple. The command line should show whose move it is and allow you to select one of these:(1) Hit (2) Stand (3) Double Down (4) Split (5) Surrender After the game ends, the program should declare a winner and ask for a new game allocating the winnings to the bank account of your user